feat<>: the file is working as it should now
it generates a folder for each training session with the date being the name of it. The folder will have the .CSV files collected from the compressed files along with a filtered .CSV that have only x_coord, y_coord and vehicle_conf columns. 
It .GPX and .HTML file for each training session but in different folder. the first one is in the main root and the second one is in '/training_maps' folder