refactor<>: split the mobile controller into 5 files inside JS folder
the new landing point file is called "mobileController_a_app.js" which links between the other 4 files. I separated them because Segment cooperation feature will require more settings to be added, which will require separation of concern for the code.
Removed the need for midpoint in the "deltaCoordinatesFromTip()" as it was static on the initialization of the two triangles.
Created setters and getters for the global variables used between deferent modules