feat<utils/ssh>: add router function to scan for networks and filter the data
The functions scan for all discoverable networks, returns a filtered JSON for the data after replacing the unknown data from IE information with one that is known based on the data in it. There is DocString for each new function.
I don't think it is thread safe, will check this with local mode