feat<>: change the parser to have dynamic ip from config files(s)
Most of read from the parser were done with a function that would check if the key exists in the config file. In case the key doesn't exist, the parser function would return default value. I added the key in the config file (not all services had the same config file to read from) and changing the ip would be for all the values at once, depending on nano's IP