feat<TEL/menu_robot>: add utils class
it has setter and gett for segments variable. IT will be used mutual between all the classes as a global variable.

Used instance of the utils class in both AdminMenu and the manger for segment table.
The goal is to append all the data in it before saving it in the robot config or sending it through the broadcasting socket