refactor<coms/app.py>:Many changes, look at description
First: Added a multiprocessing event flag that will be used to exit the code whenever necessary. This flag is used in many other services in a segment.

Second: Removed the command forwarding thread to reduce a little bit of load in the service. Instead its function was transferred to main().

Third: The command processing takes place in the FL side (the server). We need this because the command processing has to happen on the receiver. The sender just sends what he is doing atm, and the receiver decides what to do based on that.
Also, the Head might be moving with a self-driving feature, increasing load. If it also had to process commands to send to its FL, it would probably produce commands a lot slower.

Fourth: Instead of using global variables for the messages, we use the class variables of segment_server and segment_client.

The command forwarding in main() just tries to send the values of two variables. Those variables update their internal values through the server_code() and client_code() functions.