refactor<coms/server/client.py>:Look at description
In both server and client, added new class variable that store the message that is about to be sent, and the message that was received from the other side. Those variables are used as "global" variables in the app.py. These variables are constantly updating their value, and the main() of app.py simply forwards the latest values of the variables to Pilot.

Also removed "return" statements in the recv methods, since we are using the class variables, so no need to return a value when it is stored in the class itself.