feat<SSh>: Add function to get ip address from mac
The function is used when checking connection between segments. In case a segment exist, then it can get the ip of the target segment using its mac address.

This to overcome the problem that occurs when the user add a new segment from the UI. There is no way to get the ip (as the UI returns only the wifi.name and mac.address) so i can ping the device and check for a valid connection. FetchIpFromMac class will search network and wireless config of the router and get the ip address of the networks in it. The class will return ip of the router and ip of the nano, both of them in a list