feat<coms/cmd_proc.py>:Add delay feature
The LD will turn right now, and the FL must turn after a bit of delay.
Delay = Distance/Velocity
After this delay passes the FL will start applying the steering values that it has stored while the LD was turning.
Once the queue is empty, that means that the LD is also moving straight, and no steering values were being stored.