feat<teleop/app.py>:Add brake/accelerate feature
Now the robot will fully accelerate/brake within 1 second.
Before the robot would instantly try to apply the throttle the user was inputting.
Also before it would brake instantly the moment the user let go of the controller.
Now it is safer for the user and the items on the robot.