feat<PIL/relay>: remove tcp from being used in pi uri
as this function is the only one that requires having the pi uri as tcp. I added the uri of pi in the config.ini then made it add tcp later in the function after it gets the ip from the config file.
The changes allow more than one function to take the ip then they can change on it later ( it can use tcp, ipc socket if needed)