feat<config.ini>: add missing keys
The `parse_option()` works by getting a key, data type, default value as parameter and the file for where to look in.
It looks for the key in the config file, if found, then it will use it, if not (which was the case most of times), then will use the default value passed at the function call.
The `Config` class in VEH will change the ip according to each segment and any service will read the updated value from config.ini file