feat<FOL/b_FOL.js>: wrap the switch command function in a class
As the old function could be only called on the button call and couldn't be exported. I had to wrap it in a function I can export, but the function was doing more than one process (changing the text on button and sending POST with the command) so I split it into class.