feat<ras/servos> add if condition for relays
This part was removed when i was changing the clock speed (hz) for the pi relay. I wanted to simplify it but relay control is overengineered on the pi's side. There isn't much that can be done about it unless it is necessary, which is out of the scope for this project.