feat<tel_utils> process commands from fol first
the older implementation counted that fol will go through the smoothing process, but it won't. There had to be a way to split the commands from different sources and to limit to not be smoothed.

I would have made a different socket to broadcast the state of fol and leave the current socket just for commands, but i do not want to add extra socket.

Have mercy on the CPU