feat<fol, tel> change state holder for fol
The old implementon depended on the state of the buttonto decide of fol should work or not. If the button had start following, it would send the opposite command of stopping.

I wanted to add loading command as it would tell the user when the model is loaded.

This part needed to be done from fol service itself, as TEL wouldn't know by itself when FOL is ready. I had to change the implementation of UI to make the state change on the ui depends on the backend and not the ui itself, which is better, in case a problem happens in fol, the user will know about it.

I also removed the cache saving for the current state of fol button's text and colour, as it isn't needed anymore