feat<fol_utils> change azimuth adjustment logic
I had a problem where it would send steering on right azimuth (let's say with difference 240 angles) more than it would send if the azimuth was done from the left side.

Long story short, the steering needed to fix the azimuth, would be a lot from left side, but less on right side. I don't know what was causing the problem, but I changed the logic for this one.

It isn't the best, as it doesn't have a safe zone. but was working fine