feat<fol_utils> add code for calibration process
This code isn't tested yet.

The code raises a flag to start the process of calibration, while the user is still in the safe zone. The safe zone is extended by half (as both of the red zones are clipped by half) to give more space for the user to move in it