refactor<tel_utils> change the command source priority logic
The old implementation gave priority be default to the commands that are coming from following if it is active. That made a problem when getting the commands from mobile controller UI. It didn't give it a chance to increment the speed as it was taken from following source.

I had to give priority to the source that sends a throttle and steering that isn't zero, then have a little timeout until switching for another source if the current one doesn't send command anymore