refactor<fol_utils> send fol command all the time and camera only when active
I noticed that the camera stream just goes black, after I have sent lots of pan commands. Maybe the camera has a limit of the amount of commands it can receive in a specific amount of time? No clue.

Will try to implement a thread lock and see how it will behave