refactor + doc
Added comments and some inner thoughts for the code flow.

I'm thinking if i should go back to the old implementation of calibrating the vehicle while the camera is following the user, or should i stick with the calibration flag current implementation.

The calibration flag won't work if the user keeps taking small turns around the robot. the camera will see the bars eventually.