feat<fol_utils> follow person even if they are close to the camera
I noticed it wouldn't move at all after I implemented the safety feature. the user might be close to the robot and just taking a turn around them, it would be wise to make the camera follow them, without the robot moving.

This will need extra logic implementation in the future