feat<fol> use trained model
this one is trained with augmentation for 485 epochs then with much harsher ones (angel 100 degrees) for another 60 epochs.

I did perform less than the normal yolov8 and yolov10 models, but this is because the validation is done with respect to other detection classes (the other 79 neglected classes) which effects the metric

I will try to improve the model by fine tunning it for the hyperparameters.