feat<tel> add styles for the main UI
Now I need to implement the new UI for settings menu. These changes cannot go with the pre. implementation that was done by Stephan as he just pulled the data from the backend and didn't change any in it. The idea is good to have in the first place but the implementation would take quite a while.

I think I will need to change the Jquery code to make it populate the HTML and save resources