refactor<getSSID> remove the file
as I added the utils/ssh.py fiel that would get the ssid from the router using ssh and not through terminal command. Will need to refactor this part to save it in the config.ini file then read the data from it. This is a new thing not related to the code now