feat<tel> work on the UI and split files
The idea is to have a streamlined code for index.js that is imported only from index.html. the mc js files should be only one file that is included from index.js (or mc html page) but is always loaded and gets the element only when the page is loaded. Why not just import the file when the mobile controller page is loaded? because there are elements on both mc and nc that needs to be manipulated by mc js and if i import mc js only in mc html, then it will limit the js file to manipulate DOM elements of mc html only.

The code for nc page is written in mixed language of jQuery and js so I need to control on what page to pull the look for elements and where not to