refactor<tel> change tag selector from id to class
The speed meter is used on both normal and mobile UI page. The query for the DOM element, won't work if both of the pages are opened at the same time, but will work if one of them is open.
I put in mind a case where both of the UIs are open