fix<tel> storage saving for toggle advanced view
`localStorage.getItem()` returns a string, not a boolean. When I stored "false" in localStorage, it's stored as a string "false", which is truthy when evaluated as a boolean.

To fix it, I had to parse the bool value to string.

Also the saving wasn't working fine, as the cache saves bools as strings by default, so there was no point of saving the value in bool (to keep the code consistent