a hopeless try to fix multi camera stream
Too many loggers are added here to follow the trace of code.

On the normal UI when the user opens for the first time, the stream is working fine. If the user navigates to one of the settings menu then goes back to manual mode, the stream won't show up. I have been trying to add error handling (in the next commit) and see if there is a dom element that can't be reached, but I don't see any errors.

This commit is made just to come back to in the future if I (or someone else) decided to pick up on this problem