feat<docker-compose> use the new mongodb docker image
The problem I had was with the GPG key rotation that happened with mongoDB. Sometimes, I would ge tthe new key when i make a new release and others, I wouldn't. As this problem took place more than one time, I decided to work on it and change the docker image. I decided to unify the docker compose command for the docker image to only include the path to dockerFile and the shared volumes, while the rest will be done inside the dockerFiel itself (I'm planning to do the same with all other services.