feat<doc/mongodb> used a new docker image based on ubuntu focal 20
I changed he old implementation as it depended on having mongo as a base docker image. When i was looking for a solution to the GPG key problem, I can into this comment https://stackoverflow.com/a/56200257/6396830 which says it isn't recommended to have both mongo and python on the same docker image (which is the needed case here) and if so, you would need to start with a an ubuntu base (or per needed), and I took it seriously. I managed to make the docker work before with this implementation https://github.com/technologiestiftung/otc-jetson-nano-docker/blob/master/docker-compose.test.yml but it didn't have python inside of it (and also was running the mongo image from the docker compose file which was like hmmmm ok pass)