feat<tel> show image in the logbox table
The old implementation would show img then take the src from it and show it on a canvas. The canvas was used to show the lines. I removed the image and showed the img directly on the canvas

I would say this is the second resource saving that I do after removing the third stream of the non-selected camera (that was used in the overlay when you click on enlarge the second stream)