refactor<tel>: change path_render from object literal to be a class
I was testing inf and the code didn't work so i went through it to refactor some parts. The goal is to change all the code instances  to use classes instead of the object literals.

Why object literals were used in the first place? I assume because there were no import and exports in the code and all the files were imported in index directly. I changed the import to have it only from index.js and router, all the other .js files are imported into these two files