feat<tel> add features styles with css and not js
The previous impelemtation had each features styles added via js. These were lots of styles to add in each feature. I opted for another solution which is with css, which is what css is used for. It will make it easier to make the styles in js for the index page later. For now, the styles are working fine and I will polish some touches in it