New docker image+ gpio switching
I wasn't planning to start the process of common folder shared between the services here. This is just for now.

I decided to make a new docker image because the existing one was running python 3.5, which i couldn't do string interpolation in it. As far as i know, this was the only service (on both nano and pi) that had python version older than 3.7.

I don't know why it was using debian instead of ubuntu, but I think it was made on early implementation when the GPIO pins on the pi were being used (you can see on the archived folder).

The current change with the BYODR_UTILS folder will be removed, and i will use this one as the base for other nano and when I implement the common folder CLI shell script