feat<servos> revert back to old pyvesc build from docker
it turns out that the installation coded the current firmware of the vescs in pyvesc while installing it. I don't know how it would get the firmware of a connected USB before it even starts the pi to see what is connected to it. Instead of trying to adopt my solution to work with the new implementation of set Duty cycle, I decided to go back to the old solution. I tried to change and fetch the firmware from inside, but couldn't do it. I won't waste more time on it. Maybe i will return later for it