docker<stream> create new docker image for camera streams service on pi
I was trying to push a new image on the pi, after changing the directories. I image couldn't build and showed that all the system dependencies links are not reachable.

I went into the base image (from mmastrac) and found out that it is 4 years old and the encoding used in it is deprecated by raspi official github repo. I decided to create a new one from the bottom up with an upgraded python version (3.9) and upgraded debian release (I left the option to use an older release)

I will wrap this docker image later into a build image on mwlvdev account