refactor<tel/mjpeg>
Made it better in terms of  variables/functions and classes names. Also added proper comments to the functions and classes.

Not having loose functions and they should be wrapped in classes, if there are some functions that are used more than once, they are wrapped in a mjpeg utils classe.

Use DRY method when rewriting the code.

I will move these files into another folder with a readme file to explain them better