refactor<tel=>veh> move the robot config from tel to veh
When I was working on this feature previously, VEH had python 2.7. Later along the dev process, I made a new docker image with a higher python image.

As VEH is dealing with the configuration file, it makes sense to have it also doing the robot configuration file and populate it