feat<tel> add robot train UI components
Added the new robot ui menu along with the needed navigation points that are needed to make it work. Added also styles for the new components and got Massy's feedback on how the layout should be. Will implement these points later when I get to finish the implementation of connection