feat<ssh> change the connection class to use UCI instead of directly changing config files
Teltonika has a UCI (unified command interface) that can be used to make changes to the config files (wireless, network and firewall) without the need to change the files directly. My older appraoch was changing the files directly and didn't use the UCI.

I opted to partially use the UCI in most of the connections but not in all, as being done in the firewall configs. I couldn't use add_list to work on the same line, so instead i used my old approach when dealing with the firewall file on both routers

Has been tested on Earl and Frank back and forth and working fine